Move `class Scope` from composite_nodes to internal_nodes
It was accidentally moved over by #5605.
