+++ csrc/ir/composite_nodes.h
@@ -7,8 +7,6 @@
 // clang-format on
 #pragma once
 
-#include <list>
-
 #include <exceptions.h>
 #include <fusion.h>
 #include <ir/base_nodes.h>
@@ -221,73 +219,6 @@ class SdpaFwdOp : public Expr {
   std::vector<PolymorphicValue> evaluate(
       const ExpressionEvaluator& ee,
       const std::vector<PolymorphicValue>& inputs) const override;
-};
-
-class Scope {
- public:
-  using ExprList = std::list<Expr*>;
-  using Iterator = ExprList::const_iterator;
-
-  explicit Scope(Expr* owner) : owner_(owner) {}
-
-  std::string toString(int indent_size = 0) const;
-
-  const ExprList& exprs() const {
-    return exprs_;
-  }
-
-  // Used only by MultiDeviceExecutor. Should generally be avoided in favor of
-  // other modifying methods.
-  ExprList& mutableExprs() {
-    return exprs_;
-  }
-
-  Expr* front() const {
-    NVF_ERROR(
-        !exprs_.empty(), "Attempting to access the front of an empty Scope");
-    return exprs_.front();
-  }
-
-  Expr* back() const {
-    NVF_ERROR(
-        !exprs_.empty(), "Attempting to access the back of an empty Scope");
-    return exprs_.back();
-  }
-
-  bool empty() const {
-    return exprs_.empty();
-  }
-
-  int64_t size() const {
-    return std::ssize(exprs_);
-  }
-
-  Iterator insert(Iterator pos, Expr* expr);
-
-  Iterator pushBack(Expr* e) {
-    return insert(exprs_.end(), e);
-  }
-
-  void clear();
-
-  Expr* owner() const {
-    return owner_;
-  }
-
-  // The following methods perform linear searches over exprs_. Use them only
-  // when necessary, as they do not scale well with large scopes.
-  Iterator insert_before(Expr* ref, Expr* expr);
-  Iterator insert_after(Expr* ref, Expr* expr);
-  void erase(Expr* ref);
-  bool contains(Expr* expr) const;
-
- private:
-  void erase(Iterator pos);
-
-  ExprList exprs_;
-
-  //! Owner exprssion of this scope, e.g., IfThenElse
-  Expr* owner_ = nullptr;
 };
 
 // SDPA bwd node with same functionality
